+++ modules/text_completion/src/cpp/FileCompleter.cpp
@@ -52,7 +52,7 @@ splitpath(std::wstring& prefix, std::wstring& path, std::wstring& fname)
     for (size_t k = 0; k < prefix.size(); ++k) {
 #ifdef _MSC_VER
         if ((prefix[k] == DIR_SEPARATOR_OTHERS) || (prefix[k] == DIR_SEPARATOR_WINDOWS)) {
-            lastslash_pos = k;
+            lastslash_pos = (int)k;
         }
 #else
         if (prefix[k] == DIR_SEPARATOR_OTHERS) {
